requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
DateTime==4.3
db-dtypes==0.3.1
exchange-calendars==3.5.1
Flask==2.0.2
fonttools==4.28.5
google-api-core==2.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==2.32.0
google-cloud-bigquery-storage==2.11.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
MarkupSafe==2.0.1
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.1
oauthlib==3.1.1
packaging==21.3
pandas==1.4.0
pandas-gbq==0.17.0
pandas-market-calendars==3.2
Pillow==9.0.0
proto-plus==1.19.8
protobuf==3.19.3
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.3.0
pyluach==1.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
toolz==0.11.2
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zope.interface==5.4.0
yfinance==0.1.66
pandas-datareader==0.10.0